Add a lint for native_functions.yaml (#56059)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/56059

The lint doesn't do very much, mostly it enforces that indentation
is consistent.  The real point of the lint is to just make sure
that we can still do surgery on codemod with tools like ruamel,
by reusing the configuration in this script.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

Test Plan: Imported from OSS

Reviewed By: walterddr

Differential Revision: D27774590

Pulled By: ezyang

fbshipit-source-id: c26bc6c95a478bd9b86387b18de7e906e7d13193