[pyper] allow static runtime on and glow on simultaneously (#57972)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/57972

Allow static runtime to be on when glow is on. This should be fine as long as glow AOT has already been run.

Test Plan: Test on replayer with remote_other net. D28291326 fixes remaining issue removing loops from the remote_other model. Need to test on regenerated model.

Reviewed By: hlu1

Differential Revision: D28275514

fbshipit-source-id: ee78972660dfdc3fcfb9af2cf7ebb19ee745a4f1