[nnc][scripts] Add a script for bisecting the TE fuser pass (#58357)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/58357

Finding a miscompilation in a large program can be tedious; this
script automates the process of bisecting based on the number of fused
instructions.  Since fusing aten::cat without the corresponding
prim::ListConstruct will cause an assertion failure, we treat that case as a
"skip" and ignore it for the purpose of bisection.
ghstack-source-id: 129079484

Test Plan:
Tried it on some failing testcases, plus I wrote a simple bash
script to simulate "failure" and "skip" and verified a few different cases.

Reviewed By: huiguoo

Differential Revision: D28463808

fbshipit-source-id: 64836f1d37a573549179410316ea7168e3dc1f23